Removed old functions and resolvers that were being used until refactor was complete. No longer needed for production.
